Replace Net::FTP with a call to wget
I have been getting weird timeouts from Net::FTP. This tries to get
around the problem with wget, which does not give timeouts, at least
when called directly from the command line.

The downside is that we loose the ability to log properly into the
FTP server, but that can be fixed later.